Re-add missing generic type argument for Java 7
Older React Native versions use a version of the Android Gradle Plugin
that does not yet default to Java 8, which means the type argument
cannot be inferred automatically. This resulted in a compile error on RN
0.64 (as reported). The solution is to re-add the type argument, and
also set compile options to Java 7 in order to avoid these type of
issues in the future.